html/semantics/scripting-1/the-script-element/scriptexecute-events/018.html

<!DOCTYPE html>
<html>
  <title>no onbeforescriptexecute for empty script</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <script>
    targets = [];
    var t = async_test();
    document.addEventListener("beforescriptexecute",
      function(e) {
       targets.push(e.target);
    }, false);
 </script>
 <script></script>
 <script>
   t.step(function(){assert_array_equals(targets, [document.getElementsByTagName("script")[4]])});
   t.done();
 </script>
